backstage: o-header, improve inconsistent labelling
Inconsistent Labels

This commit improves consistency of labels. Previously: The menu button provided for users in the header, has been labelled as ‘open side navigation menu’. Once the menu is opened, the close button is labelled as ‘Close drawer menu’. This can be confusing for users of screen reading assistive technologies, as they are not informed that both components refer to the same menu. This can be increasingly difficult to discern when navigating the page out of context using the forms list using the names of forms alone.
https://github.com/Financial-Times/origami/issues/901

Visual Navigation Headings

Text had been visually styled as headings, without being programmatically marked up as such for users of screen reading assistive technologies. 

This commit takes a step in that direction by breaking multiple navigation lists up. More work will follow in a future commit.

https://github.com/Financial-Times/origami/issues/910