feat: o-normalise, add double focus ring by default for inputs.
Our default focus outline has been designed to have sufficient contrast
against paper and slate backgrounds. It has poor constrast in some cases
which require extra attention – for example against wheat. In these cases,
which are not so common, we usually set a new focus colour with more
contrast.

However one repeated and common case where our focus styles are not
obvious is when the contrast between the focus outline and the focused
element is low e.g. on buttons, checkboxes, etc. The focus outline must
contrast against the page background and the focused element.

To improve this we have a new focus style which uses 2 outline colours,
1 dark and 1 light, which is intended for inputs including buttons.
This ensures there is always a contrasting outline. The `box-shadow`
attribute is used to achieve this which does not look good for
multi-line text, therefore we stick with the old focus style by default.
Example here:
https://adhocteam.us/2022/02/08/creating-focus-style-for-themable-design-system/#focus-style-color-exploration

New mixins will allow us to apply the new focus style as needed:
- `oNormaliseFocusApply` - to apply given styles using `:focus-visible` with a `:focus` fallback
- `oNormaliseFocusContent` – focus styles for dark elements on light backgrounds
- `oNormaliseFocusContentInverse` - focus styles for light elements on dark backgrounds
- `oNormaliseFocusContentForElementColour` - focus styles given an element's background colour, useful for dynamic themes
- `oNormaliseFocusUnsetContent` - to remove focus styles

We no longer use the `:focus-visible` polyfill as there is now
wide support for `:focus-visible` and the polyfill adds complexity
and specificity. We provide a fallback using `:focus` for browsers
which do not support `:focus-visible`.