fix: o-footer, allow links to wrap rather than overflow with ellipses
At some breakpoints links within o-footer are partially hidden when
they overflow. This also means there are hidden footer links when
zoomed or when a custom text spacing is applied – an accessibility
issue.

Instead allow links to wrap, so no link copy is ever hidden.

close: https://github.com/Financial-Times/origami/issues/931

When this happens it is more difficult to skim links left to right,
as links in each column may not be aligned. This isn't such an issue
since links are categorised in columns. We may choose to undertake
a broader change in the future to reduce this by:
1. Restoring the "More from the FT Group" link from its current
position, in its own column, to sit on a new row spanning the
footer's width.
2. Updating footer link columns to use just the space they need,
rather than taking up a specifc number of columns based on our
grid system.
The above requires more input from the design team.