feat: ft-concept-button, add inverse monochrome variant
This will be used first within o-topper on the article page, on white backgrounds. Typically Origami components use "inverse" themes to indicate a white foreground, for use on a dark background. However in this case ft-concept-button has an "inverse" theme already which is intended for use on claret backgrounds, and uses a claret text colour when clicked. We want an inverse theme that is also monochrome...