feat: o-forms, allow submit events to be intercepted
By default o-forms listens to submit events. It validates a form and,
if it is valid, will then call submit on the form element. This means
projects which use o-forms cannot prevent a valid form from submitting.

This may be desirable to add extra steps prior to submission. For
example to load a captcha (apparently this should happen on submit?)
or to display a final confirmation.

To achieve this option:

This PR introduces the `oForms.submit` event. This is fired by o-forms
when its submit event handler has finished form validation, just before
submitting the form. A user may listen for this event and prevent
propagation to stop the form from submitting. As this would rely on
the project's event listener running first, however, the there is now
also an option (`preventSubmit`) to ensure the form is not submitted by
`o-forms`.

E.g:
```js
import oForms from '@financial-times/o-forms';

// Initialise a form. Note `oForms.init()` could also be used.
const formElement = document.forms[0];
new oForms(formElement, {
	preventSubmit: true
});

// Listen to the form submit event to handle submission manually.
formElement.addEventListener('oForms.submit', (e) => {
	console.log(`A user would like to submit a form, it is ${e.detail.valid ? 'valid' : 'invalid'}.`);
	if(e.detail.valid) {
		e.detail.instance.form.submit();
	}
});
```