fix: o-typography, position visually hiddent anchor content relatively [OR-148] (#1047)
* fix: o-typography, position hidden anchor content relative to the anchor.

Anchor elements which open in a new context (tab/window) have some hidden copy in a pseudo element to let screen reader users know. This is visually hidden with absolute positioning and clipping. This may cause an issue if it has an ancestor with hidden overflow between it and whatever element on the page is its containing block. To avoid this, we set `position: relative` on the anchor element to ensure it is always its containing block and so any ancestor with hidden overflow will clip it.