backstage: o-multi-select, update spacing, remove o-forms peer-dependency
Pre-release, o-forms was used as a depdendency in o-multi-select Sass.
The o-forms Sass variables used started with an underscore, indicating
private variables:
https://sass-lang.com/documentation/at-rules/use#private-members

To use these in o-multi-select we would have to make them public.
I chose to remove o-forms as a dependency instead. We will have to keep
them in sync manually, remembering to update o-multi-select if we ever
change the spacing in o-forms. The impact of forgetting is low, and it would
be easy to correct. I opted for this as:
- The o-forms variables are poorly named. They are not based on use-case.
A change to the o-forms variables will not definitely mean we want the same
to apply to o-multi-select.
- A dependency complicates Origami. A major release of o-forms would mean we
need to do a major release of o-multi-select, even if the changes
in o-forms did not impact o-multi-select. This significantly
increases the cost and complixity of rolling out changes to our
users as more users would be affected, and they would have more work
to do.

At the same time the spacing consistency has been improved. Selections
are now aligned with the comobox list, as per the design. To achieve
this I had to reduce the padding to 8px. The design uses a mix of 8/9px.

Slightly weirdly, this commit also sets the height of o-multi-select
using `oSpacingByIncrement` in the same way as o-forms. This is
for users who have enabled relative spacing in Origami, `oSpacingByIncrement`
will output `rem` values in this case.