backstage: o-multi-select, update destroy method
Do not remove the root o-multi-select element. Ideally the destroy
method should revert any changes made by the component's JS.
https://origami.ft.com/specification/v1/components/javascript/#destruction
However this is not always necessary or practical, as is the case here.
Instead we just remove global event listeners (this commit also fixes
that by passing the event handler to `removeEventListener` correctly,
previously it did not work as the actually listener was a function which
wrapped the handler).