feat: option to keep related content visible in header-services
By default, any related content in o-header-services shifts into the
drawer on small screens.

In spark, we use this space for the New Content button, which we would
like to retain visible in the nav on small screens, as it is a critical
feature.

This change introduces a new classname on the related content element,
which prevents it being shifted into the drawer and keeps it visible at
all breakpoints.

It also adds a `relatedContentAlwaysVisible` prop to the TSX template.