backstage: update Percy baseline on release commit to main (#1268)
We are running Percy on any commit to main, including release commits, as the release commit message includes a PR reference e.g `chore: release main (#1266)`. This means we run screenshots on all components twice when merging and releasing a change. Instead, use `includes` to detect release commits more readily. And instead of running Percy on any commit to main except release commits, do the opposite. We only need to update the baseline if we are releasing some change.