feat!: Upgrade to node 18, React 18 & storybook 7
This may be a breaking change for all stories & components, especially on older versions of Node.

This change was necessary as it's v. difficult to use React18 on other repos, which blocks a few
modern react features necessary for speeding up the FE

Still a WIP until I can figure out the purpose of the storybook decorators, particularly withHtml, as
well as verifying if the figma plugin is working correctly