feat!: Upgrade to node 18, React 18 & storybook 7
This may be a breaking change for all stories & components, especially on older versions of Node.

This change was necessary as it's v. difficult to use React18 on other
repos, which blocks a few modern react features necessary for speeding up the FE

I've removed the withHTML addon, but not deleted it. Notsure what it's
purpose is, but will ask in the storybook channel