feat!: Upgrade to node 18, React 18 & storybook 7
This may be a breaking change for all stories & components, especially on older versions of Node.

This change was necessary as it's v. difficult to use React18 on other repos, which blocks a few
modern react features necessary for speeding up the FE

I've removed all the addons from storybook as I can't understand what they're doing, or whether they're
still needed. I will mention this in slack and on the PR and hope someone can reassure me or help me
restore them