fix: Remove "engines.node" from o3 components
Origami components do not require node except for development, so we decided to remove engines.node to avoid warnings when users are building components with another node version. The origami repo is a mono repo, so we can rely on the root engines.node for development. Components do require a flat dependency tree, enforced through peer dependencies, so we change engines.npm to a permissive >7.