fix: Rename sustainable views brand, remove specialist brand.
Since specialist isn't a brand in itself, but a general grouping of brands, it might make it harder for us implement any sub-brands of a specialist title. 

Our current logic Github side doesn't handle the space :cowboy-with-grimacing-face-with-smiling-eyes: I can fix that, but wonder if it'll translate the same in Figma variables etc. Maybe best to be explicit with a dash