fix: allow for multiselect options with whitespace
Currently we use a string label to create the multiselect option
elements. When one of these labels has a whitespace character, it cannot
be removed, as querySelector doesn't find the element with the
whitespace in the ID.

This change adds a new optional "key" to the TSX template (allowing
users of the TSX to have a different value to the label), and ensure the
value is used for creating the ID (rather than the label).