fix: o-multi-select, use textContent rather than innerText (#1388)
This is to improve compatibility with JSDOM for users' tests. It shouldn't change the functionality of o-multi-select. However, they're not equivalent APIs, and it would be reasonable to use `innerText` or other web API which depends on CSS layout. We recommend users test components in a real browser, and can't guarantee changes to support executing component code in JSDOM.