fix: remove selector-no-qualifying-type stylelint rule
This is helpful to think in terms of BEM, and avoid breaking markup changes – we should prefer classes which can be applied to any element. However we override it regularly, there are cases we want to do this, such as changing forced colours based on whether a button component is an `a` or `button` tag.