backstage: Refactor Token Studio build scripts. (#1430)
- Update API to encapsulate Token Studio configuration.
- Reduce duplicated logic.

There is one `todo` around appending an `_` to private CSS custom properties. This commit underscores anything beginning with `o3-button` specifically as `isSource` is always true. We got away with it before as the transform was applied only on component design tokens, but we should be able to make selected component tokens public, any any other design token of our choosing private.

We intend to work with our designers to see if we can treat `o3-button` design tokens `as source`. I think that will be a good thing, as it won't create Figma Variables or styles for source tokens. We can then update our build script to use `isSource` rather than match on `o3-button` specifically.
https://docs.tokens.studio/themes/token-sets#treat-as-source