fix: Replace o3-button theme class with data attribute
Often whole sections of a page are themed. Instead of apply a theme class
to every element/component within the section, we would like to be able
to apply a theme on a parent element and have all children use that theme.
This is similar to how o3 component brands are applied.

This commit aims to achieve that for o3-button. Future components
should follow the same pattern.

E.g. A dummy example...

Existing o2 components, branded with inverse theme:

```html
<body>
	<button class="o-buttons o-buttons--primary o-button--professional-inverse">
		Old inverse professional button
	</button>

	<p class="o-typography-body o-typography-professional o-typography-inverse">
      Here is a <a href="#" class="o-typography-link">standard link</a>.
    </p>
</body>
```

Becomes:

```html
<body data-o3-brand="professional" data-o3-theme="inverse">
	<button class="o3-button o3-button--primary">
		New inverse professional button
	</button>

	 <p class="o3-typography-body">
         Here is a <a href="#" class="o3-typography-link">standard link</a>.
     </p>
</body>
```

However, sometimes, we do want a theme to apply specifically to an
individual element. This is also supported and takes precedence.
Another dummy example:

```html
<body data-o3-brand="core">

<section data-o3-theme="inverse">
	<button class="o3-button o3-button--primary">
        An inverse button
    </button>
    <p class="o3-typography-body">
         Inverse body copy.
    </p>
</section>

<button class="o3-button o3-button--primary" data-o3-theme="mono">
    A mono button
</button>

</body>
```