fix: always force font anti-aliasing
This isn't standard:
https://developer.mozilla.org/en-US/docs/Web/CSS/font-smooth

And is probably not something we should do:
https://www.zachleat.com/web/font-smooth/

But it's something we already do and if we don't, font weights
look lighter in Figma than on the web. And that's fine and good.
But it's confusing for designers, and I can't die on this hill.

Removes the undocumented utility class that it replaces.