fix: Include icons in branded CSS files.
Do not publish a seperate `icons.css` file. This will allow us to
brand icons if needed without a major change. It also makes foundations
easier to use as there is 1 file to include. As the previous behaviour
was not documented and adoption is low, I believe we can treat this as
a patch release.

This commit also documents colours and spacing, as well as adding
an icons example.