feat: Allow some component tokens at the brand root.
By default component tokens are scoped in CSS to a component class:

```css
[data-o3-brand="core"] .o3-component-name {}
```

This is helpful because:
1. In-browser Dev Tools are easier to use.
2. Enforce they are used by the component only, especially where private.

There is no other real benefit to scoping as all CSS Custom Properties which are created via design tokens are prefixed with the component name.

However that means the root `.o3-component-name` has to be applied on or as a partent to all component elements. This is not a great experience for components with many utility classes such as `o3-typography`.

E.g. the extra "o3-typography" makes utility classes verbose:
```html
<span class="o3-typography o3-typography-bold"></span>
```

This commit introduces an opt-out, so we can move component CSS custom properties to the brand root rather than scope under the component class. We may find we want to simplify in the future and move all component CSS Custom Properties to the brand root. Let us see how it goes as it's easy to make the selector less specific, but hard to make it more specific.