backstage: Post-process script to remove links to missing headings.
To avoid duplicating content, we want to write brand specific
sections of mdx using .astro components to render portions of
mdx only for select brands. However Starlight finds headings
by inspecting mdx before this happens. This means content is hidden
correctly, but there are links within table of contents components
(sidebar, in page links) which show links to headings which do not
exist on branded pages.

This post-process script removes heading links from table of content
components where the heading does not exist on the page. It is a little
fragile, as it depends on Strlight markup not changing. We decided as
a team this is a reasonable approach for the medium term. It was our
prefered approach compared with maintaining a fork, or other much more
time consuming customisations.

We anticipate moving away from Starlight in 2025 to re-brand Origami
to the FT Design System.