backstage: Patch Starlight to pass "brand" to mdx content.
We "brand" content using Starlight's internationalisation feature.
We duplicate lots of content currently as we need to make small
content changes per brand (language), e.g. to show a "mono" theme
for the button component only for the "core" brand. If we
pass the brand (language) to mdx we can avoid duplicating content
(super hard to manage) and instead show this section on condition
of the current brand. We intent to make an upstream PR to pass
down this data. In the meantime, we patch.