fix!: Support Coral's illegal comments report feature for anonymous users (#1689)
DSA compliance.

The Coral link to report illegal comments works by reopening the same page in a new window, with some added query string items. However, if someone is a first-click free user this will result in the barrier page loading instead. This means that anonymous users can't use the tool to report illegal comments, which is a condition of the DSA.

This change intercepts the clicks on this link, alters the path to a route that is not behind the paywall, then opens that url in a new window.

https://financialtimes.atlassian.net/jira/software/c/projects/CI/boards/1653?assignee=5be57e310e625003a46e9693&selectedIssue=CI-2234