docs: Add basic whitelabel support to the website.
This shows the drop cap and summary, which are not yet used
by any brands but will be for future Specialist Titles.

More work is required including checking links to figma &
storybook – some are set, some aren't. This may benefit
from dedicated time to make Figma/Storybook links easier
to define and manage – could we define a Figma location
once for Storybook's panel and the site?