fix: Remove prefers-reduced-motion from normalise.css (#1801)
The `prefers-reduced-motion` setting is binary and does not give users
much control or choice. Prefering reduced motion does not mean the
user wants no animation or transition at all. It should be used on a
case by case basis. For example to stop the largest and most potentially
problematic cases (e.g. parallax). By defaulting to to no animation or
transition with `!important` we make it difficult to undo.

```css
/* Reduce motion */
@media (prefers-reduced-motion: reduce) {
	*,
	*:before,
	*:after {
		animation-duration: 0.001s !important; /* stylelint-disable-line declaration-no-important */
		transition-duration: 0.001s !important; /* stylelint-disable-line declaration-no-important */
		animation-iteration-count: 1 !important; /* stylelint-disable-line declaration-no-important */
	}
}
```

https://css-tricks.com/nuking-motion-with-prefers-reduced-motion/