fix: change lineheights to dimensions
Style Dictionary does not know how to correctly translate a px value with no specific type identifier to rem/px.

E.g in Token Studio, a lineheights value could be 32, Style dictionary reads this as a unitless scale, and will try and set lineheight to be 32x the font size.. not what we want.

Instead, we define lineheights as dimension types, Style Dictionary knows to add `px` to these values and subsequently convert to `rem`.

Need to confirm this does not adversely affect Figma.