fix: (o-header) subnav scroll behaviour
There is a bug in the way we calculate how far to scroll in order to make
 sure the selected item is in view in the SubNav.

We were previously calculating the scroll position by first checking how
far off the item is out of view via `currentSelectionEnd - wrapperWidth`.
If this was less than the wrapper width, we would only scroll by half
the width of the wrapper, which is not the right thing to do, especially
if the item is only slightly out of view.

The fix includes:

* Removing the unnecessary scroll by half the wrapper width logic.

* Calculating the right edge of the selected item `currentSelectionEnd`
relative to the wrapper more accurately by taking into consideration
the wrapper's left boundary (rather than just checking the wrapper width)
via:  `currentSelectionRect.right - wrapperRect.left`.

* Calculating the scroll distance more accurately by including a fixed
buffer distance `RIGHT_ARROW_BUTTON_OFFSET` for the right arrow button.
This is required otherwise, the selected item text will be slightly
cropped off by the button.