feat: o-header subnav primary styles
The App has a requirement to have a sub navigation where link items
have 'primary' styles and are capitalised with font weight 600. The
navigation also requires the background image to use `paper` colour
instead of the current default `white-60`. Esentially we want the sub
navigation to look like a primary navigation but still have the
scroll behaviour of the sub navigation.

This introduces `o-header__subnav--primary` modifier which will provide
the required style modifications for background colour and text style.

Additionally since the text styles are now being using in two places, I've
extracted these out to a `_oHeaderPrimaryLink` mixin.