chore: Make base type tokens tokens.
We do not want to export our typography scale in Figma for end users, as we have specific usecases now. It is a bit of a pain to exclude the generated Figma variables or styles from publihsing. To support migration efforts we want it to still be available in the engineering token set.