chore: Refactor the Preview component
A new PreviewPane component is responsible for the UI of the preview
interface. This makes it more flexible. It can now be used for smaller
previews, such as for utility classes which aren't a component and
do not have JSX.

The Preview component retains logic to render a JSX component,
fetch its JSX from a given file path for a code snippet, and
generates a HTML code snippet from there.