chore: Add further new type usecases
Skip link usescases. It adds an underline to the body-base style, but not the colours or interactive states we'd associate with a link. We don't have a similar style for a link within body-lg which suggests we don't expect links in those contexts? From an engineering perspective I'd have body-base or body-lg classes, and any link within would have colour and an underline for the different states (hover, etc). Unsure how to represent this in Figma helpfully for designers except for providing a link component, which doesn't help with applying a link style within a body of text. Waiting for feedback from design.