chore: correct editorial body token use
Increases mobile font size. Adds `strong` and `em` styling to direct children.

The token `body-content-highlight` font weight is coming through as text `Bold`.
I have asked the design team to take a look. `body-highlight` is correct, it's coming
through as a reference to an `o3-font-weight` value.

Moving on in the meantime.