fix: Remove `o3-editorial-typography-byline-location`. (#2012)
This has the potential to be a visually breaking change where `o3-editorial-typography-byline-location` is used outside of `o3-editorial-typography-byline`. We're happy with that given what we know about `o3-editorial-typography` use, vs. keeping that class around deprecated.

From Kara:
if we look at the example in the docs, "Joe Doe in London", it's marked up with Joe Doe as author and in London as location. but not all the text in the byline that isn't an author is a location, you can have bylines that are like "Dan McCrum in London and Antoine Gara in New York", where the "and" isn't part of a location. in the content pipeline data, this is structured as `author: Dan McCrum, text: in London and, author: Antoine Gara, text: in New York`, and the text nodes don't get any wrapper elements.