fix: Prevent an error if no sticky search button is used.
subs.ft.com heavily modifies the header. It uses a sticky header without search interface.

Generally we don't support this level of customisation for components, and would expect things could break in minor/patch releases without notice. I'd have recommended creating a custom component here from scratch. We might want to consider an Origami header component that centres a logo and allows for custom items left/right, but it's unlikely to make it on our roadmap soon.
https://origami.ft.com/getting-started/customisation/

Despite that, this is a small issue, so let us be helpful and check for the presence of search and fail gracefully. We have recommended to pin to the patch release to reduce risk of future changes breaking unsupported customisations.