fix: Remove standfirst hover and focus colour changes (#2232)
Some standfirst hover and focus colour changes are so minimal that the difference is barely visible.
We have received similar DAC issues about not enough contrast with standfirst hover & focus often.

We decided to remove these styles because:
- In some teasers, the colour change is almost impossible to see.
- The main link is the headline, not the standfirst.
- ft.com teaser's standfirst are opted out from tabbing (which means no focus situation).

For these reasons, the hover and focus colour changes is not so important for standfirsts.

Co-authored-by: Ben Freshwater <ben.freshwater@ft.com>